requirements.txt
pandas==2.2.2
requests==2.32.3
shodan==1.31.0
python-dotenv==1.0.1
streamlit==1.43.2
plotly==5.24.1
fpdf2==2.8.1
